requirements.txt
annotated-types==0.7.0
anyio==4.5.2
click==8.1.8
dependency-injector==4.46.0
exceptiongroup==1.2.2
fastapi==0.115.12
h11==0.14.0
idna==3.10
pydantic==2.10.6
pydantic_core==2.27.2
PyJWT==2.9.0
python-multipart==0.0.20
PyYAML==6.0.2
sniffio==1.3.1
SQLAlchemy==2.0.40
starlette==0.44.0
typing_extensions==4.13.1
uvicorn==0.33.0
